Update README.md
Working with the cout Statement

In the previous week's lab we saw how the cout statement can be used to display text output. To get started, copy your code from Lab 1.1 into the editor panel.

Exercise 1

Fill in code so that the program will do the following:

Write your first and last name on one line.
Write your sign of the zodiac on the next line.
Write your favorite color on the next line.
Write your favorite animal on the next line.
Remember that to output a string literal, such as "Hello", you must use quotes.

Run the program.

Example Output:

Hector Smith
Taurus
Blue
Rabbits

Exercise 2

Change the program so that a blank line separates your zodiac sign from your favorite color.

Hector Smith
Taurus

Blue
Rabbits

Exercise 3

Change the program so that the following (but with your name and info) is printed, including asterisks. Try to get the spacing for the right column to align similar to the the example.

************
Programmer:      Hector Smith
Zodiac Sign:       Taurus

Favorite Color:    Blue
Favorite Animal: Rabbit
***********

Run the program and verify it's correct. Submit the code when you are finished with Exercise 3. You do NOT have to submit code individually for Exercise 1 and 2.